fix(vdlserver): use parent dirs in ReadWritePaths to avoid namespace failure
systemd applies ReadWritePaths before ExecStartPre runs, so subdirs that
don't yet exist cause NAMESPACE errors. Use /tmp and ~/configs (which always
exist) as the sandbox paths instead.
