+++ pkgs/nixos/pc-base.nix
@@ -245,7 +245,7 @@ in
     ../python-packages/flasks/anix-upgrade-ui/module.nix
     ../python-packages/flasks/tester/module.nix
     ../python-packages/flasks/tasks_ui/module.nix
-    ../python-packages/flasks/ttvd/module.nix
+    ../python-packages/flasks/videodl/module.nix
     (
       let
         # Pinned to d4f7c8220fa5 (before PR #485 which added pre-switch-checks.nix,
@@ -538,7 +538,7 @@ in
       rcrsync = machine-rcrsync;
     };
 
-    services.ttvdserver = {
+    services.vdlserver = {
       enable = cfg.isATS;
     };
