+++ pkgs/nixos/service-ports.nix
@@ -28,5 +28,5 @@
   tester = 5757;
   anix_upgrade_ui = 5858;
   tasks_ui = 5959;
-  ttvd = 6060;
+  videodl = 6060;
 }
