+++ pkgs/python-packages/flasks/videodl/setup.py
@@ -1,0 +1,9 @@
+from setuptools import setup
+setup(
+    name='vdlserver',
+    version='0.0.0',
+    py_modules=['vdlserver'],
+    entry_points={
+        'console_scripts': ['vdlserver = vdlserver:run']
+    },
+)
