+++ pkgs/python-packages/flasks/videodl/templates/index.html
@@ -4,7 +4,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>TTVD</title>
+    <title>Video Downloader</title>
     <style>
         * {
             box-sizing: border-box;
@@ -414,19 +414,19 @@
     </div>
 
     <div class="container">
-        <h1>TTVD</h1>
+        <h1>Video Downloader</h1>
 
         <!-- Status message -->
         <div class="status-message" id="statusMessage"></div>
 
         <!-- URL fetch card -->
         <div class="fetch-card">
-            <label for="urlInput">TikTok URL</label>
+            <label for="urlInput">Video URL</label>
             <div class="url-row">
                 <input
                     type="url"
                     id="urlInput"
-                    placeholder="https://www.tiktok.com/@user/video/..."
+                    placeholder="https://www.youtube.com/watch?v=... or TikTok, etc."
                     autocomplete="off"
                     autocorrect="off"
                     spellcheck="false"
@@ -511,7 +511,7 @@
         function doFetch() {
             const url = document.getElementById('urlInput').value.trim();
             if (!url) {
-                showStatus('Please enter a TikTok URL.', true);
+                showStatus('Please enter a video URL.', true);
                 return;
             }
             clearStatus();
